Rename agregateRoots package to aggregateRoots
