+++ src/main/java/br/edu/ddd/project/usermanagement/application/usecase/CreateUserUseCase.java
@@ -5,7 +5,7 @@ import br.edu.ddd.project.usermanagement.application.dto.response.UserResponse;
 import br.edu.ddd.project.usermanagement.application.exception.UserNotFoundException;
 import br.edu.ddd.project.usermanagement.application.port.UserRepository;
 import br.edu.ddd.project.usermanagement.application.service.DomainEventPublisher;
-import br.edu.ddd.project.usermanagement.domain.model.agregateRoots.User;
+import br.edu.ddd.project.usermanagement.domain.model.aggregateRoots.User;
 import br.edu.ddd.project.usermanagement.domain.model.valueobject.*;
 import br.edu.ddd.project.usermanagement.domain.model.enums.*;
 import org.springframework.stereotype.Component;
